[simple-memo] refactor : 공통으로 사용하는 ObjectMapper 상수 추출
